Deal Better With Binder-Internal Links
Only remove links not in the allow list if they are in the left hand nav; if they're in the body of
a doc, strip the link and leave the text. Additionally, removes the bottom navigation link(s),
because they can't reliably point to a doc that's in the allow-list.